agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "LiveKit agent worker scaffold for Cozmo Voice Platform"
requires-python = ">=3.12"
dependencies = [
  "cozmo-contracts",
  "dnspython>=2.7.0",
  "livekit-agents[google,deepgram]~=1.5",
  "pymongo>=4.10.1",
  "prometheus-client>=0.21.1",
  "pydantic-settings>=2.8.1",
]

[tool.uv]
package = false

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
